Updated license as discussed in #10
