+++ bounded-core/src/main/scala/io/cafienne/bounded/aggregate/Protocol.scala
@@ -1,3 +1,6 @@
+/**
+  * Copyright (C) 2016-2018 Cafienne B.V. <https://www.cafienne.io/bounded>
+  */
 package io.cafienne.bounded.aggregate
 
 import java.time.ZonedDateTime
