+++ cargo-sample/src/test/scala/io/cafienne/bounded/cargosample/SpecConfig.scala
@@ -1,18 +1,6 @@
-// Copyright (C) 2018 the original author or authors.
-// See the LICENSE file distributed with this work for additional
-// information regarding copyright ownership.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
+/**
+  * Copyright (C) 2016-2018 Cafienne B.V. <https://www.cafienne.io/bounded>
+  */
 package io.cafienne.bounded.cargosample
 
 import com.typesafe.config.ConfigFactory
