+++ project/Dependencies.scala
@@ -1,3 +1,6 @@
+/**
+  * Copyright (C) 2016-2018 Cafienne B.V. <https://www.cafienne.io/bounded>
+  */
 import sbt._
 
 object Dependencies {
