+++ project/plugins.sbt
@@ -1,3 +1,6 @@
+/**
+  * Copyright (C) 2016-2018 Cafienne B.V. <https://www.cafienne.io/bounded>
+  */
 addSbtPlugin("com.github.gseitz" % "sbt-release" % "1.0.6")
 addSbtPlugin("com.geirsson" % "sbt-scalafmt" % "1.4.0")
 addSbtPlugin("org.scoverage" % "sbt-scoverage" % "1.5.1")
