+++ version.sbt
@@ -1,2 +1,4 @@
-
+/**
+  * Copyright (C) 2016-2018 Cafienne B.V. <https://www.cafienne.io/bounded>
+  */
 version in ThisBuild := "0.0.8"
